pyproject.toml
[project]
name = "etl-stock"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyter>=1.1.1",
    "logfire>=3.2.0",
    "pip>=24.3.1",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.0.1",
    "sqlalchemy-orm>=1.2.10",
    "sqlalchemy>=2.0.37",
    "yfinance>=0.2.52",
]
